+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the "takeabreak" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+### [0.0.4]
 
-- Initial release
+- Feature: Now the interval will only be activated if the current window state is focused.
+
+### [0.0.3]
+
+- Fix a bug that was causing the extension timer to break once the user changed a interval using any command.
+
+### [0.0.1]
+
+- Initial release of Compile Time
+
+## [Unreleased]